Code from SPIS Final Project
Color Mode: Takes a picture with the camera and then divides the image
into squares that each contain some amount of pixels. We take the
average color of pixels in each square and then attempt to go towards
the one that is closest to a chosen color.

Motion Sensing: We take two pictures and then compare the differences
between the two in a similar way to color sensing. The robot will head
towards the most different part.